+++ ydb/library/binary_json/benchmark/ya.make
@@ -12,17 +12,12 @@ IF (BENCHMARK_MAKE_LARGE_PART)
 ENDIF()
 
 SRCS(
-    read.cpp
+    write.cpp
 )
 
 PEERDIR(
     library/cpp/testing/unittest
     ydb/library/binary_json
-    ydb/library/yql/minikql/dom
-    ydb/library/yql/minikql/invoke_builtins/llvm14
-    ydb/library/yql/public/udf/service/exception_policy
-    ydb/library/yql/core/issue/protos
-    ydb/library/yql/sql/pg_dummy
 )
 
 YQL_LAST_ABI_VERSION()
